Fix most warnings in openj9.dtfj
* don't use raw types
* add missing @Override annotations
* remove redundant casts
* properly reference static features
* remove misplaced javadoc
* fix and simplify Symbol.SymbolComparator
* simplify comparator in MemoryRange
* use modern API, e.g. Collections.emptyIterator()
* suppress some unavoidable serialization warnings
* remove unused, empty collections
* fix loss of precision due to implicit casts
* remove suppression of rawtypes and unchecked warnings
* remove dead code, unused fields
* fix infinite recursion in SegmentSortException.getCause()
* add explicit constructors
* avoid using deprecated API
* remove obsolete API related to "packed"
* prefer interfaces over concrete types (List instead of ArrayList)
* mark classes final to avoid most "this-escape" warnings
* make GCIterator & SlotIterator classes generic

+++ debugtools/DDR_VM/src/com/ibm/j9ddr/vm29/pointer/helper/J9MemTagHelper.java
@@ -167,15 +167,15 @@ public class J9MemTagHelper
 		 * conditional on the extra 2 U_32s
 		 */
 		for (int i = 0; i < (J9MemTag.SIZEOF / U32.SIZEOF); i++) {
-			sum ^= slots.at(i).longValue();
+			sum ^= (int) slots.at(i).longValue();
 		}
 
 		if (J9BuildFlags.J9VM_ENV_DATA64) {
 			U32 a = new U32(UDATA.cast(tag).rightShift(32));
 			U32 b = new U32(UDATA.cast(tag).bitAnd(U32.MAX));
-			sum ^= a.longValue() ^ b.longValue();
+			sum ^= (int) a.longValue() ^ (int) b.longValue();
 		} else {
-			sum ^= tag.longValue();
+			sum ^= (int) tag.longValue();
 		}
 
 		if (sum != 0) {
